create Suspended UserWallet that wait for autoconnect
